EE-871: move pure Mint logic into separate crate
