+++ execution-engine/contracts/system/mint-token/src/contract_runtime.rs
@@ -1,11 +1,6 @@
 use contract::contract_api::runtime;
+use mint::RuntimeProvider;
 use types::{account::PublicKey, Key};
-
-pub trait RuntimeProvider {
-    fn get_caller() -> PublicKey;
-
-    fn put_key(name: &str, key: Key) -> ();
-}
 
 pub struct ContractRuntime;
