+++ execution-engine/mint/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "casperlabs-mint"
+version = "0.1.0"
+authors = ["Henry Till <henrytill@gmail.com>"]
+edition = "2018"
+
+[features]
+std = []
+
+[dependencies]
+types = { version = "0.1.0", path = "../types", package = "casperlabs-types" }
