+++ execution-engine/mint/src/runtime_provider.rs
@@ -1,0 +1,7 @@
+use types::{account::PublicKey, Key};
+
+pub trait RuntimeProvider {
+    fn get_caller() -> PublicKey;
+
+    fn put_key(name: &str, key: Key) -> ();
+}
